Age X Size models

Helper functions for implementing age x size models

all_ages(expr, fun, ...)

Arguments

expr
The expression to modify. See details.

fun
The function to apply over expr.

...
Used internally. Modifying this will likely cause models to break.

Value

A modified expression.

Details

Currently only all_ages is exported. This is a helper to indicate that all of the levels of age should be included in a given expression, rather than split out into separate levels. The fun is used to combine the results of each individual evaluation of expr.

The most common use case for all_ages will be when defining the size distribution of new recruits in an age structured model, when many F kernels are multiplied by their respective age X size distributions, and then summed. Note that using sum is almost never correct, as this will always return a single number whereas the size distribution should be a vector! Generally, use + instead.