Helpers for IPM construction

Helpers for IPM construction

define_pop_state(proto_ipm, ..., pop_vectors = list())

define_state_vars(proto_ipm, ...)

define_env_state(proto_ipm, ..., data_list)

define_hier_effs(proto_ipm, ...)

Arguments

proto_ipm
The name of object you wish to append the new kernel to.

...
Named expressions. See details for more information on their usage in each define_* function.

pop_vectors
If the population vectors are already pre-defined (i.e. are not defined by a function passed to ...), then they can be passed as a named list here.

data_list
A list of named values that correspond to constants in the formula. You do not need to specify vectors corresponding to the domains here.

Value

All define_* functions return a proto_ipm.

Details

These are helper functions to define certain types of IPM classes. It is recommended to use them within the IPM construction pipeline somewhere between add_kernels and make_ipm. The order in which they are called does not matter provided all values on the left hand side of the ... appear in the right hand side of another expression used elsewhere. Expressions where the left hand side does not appear anywhere else will be ignored during construction, and so will not be present in the returned ipm object.

Each define_* function takes dots in the same way: named expressions are captured and evaluated later in the appropriate contexts with their respective kernels. Suffix expansion is supported so that hierarchical models with year/plot/what-have-you effects do not need to be rewritten 10s or 100s of times.