Compute the standardized left and right eigenvectors via iteration

Compute the standardized left and right eigenvectors via iteration

right_ev(ipm, ...)

# S3 method for simple_di_det_ipm
right_ev(ipm, n_iterations = 100, ...)

# S3 method for general_di_det_ipm
right_ev(
  ipm,
  n_iterations = 100,
  mega_mat = NULL,
  mega_vec = NULL,
  keep_mega = FALSE,
  ...
)

left_ev(ipm, ...)

# S3 method for simple_di_det_ipm
left_ev(ipm, n_iterations = 100, ...)

# S3 method for general_di_det_ipm
left_ev(ipm, mega_mat, mega_vec, n_iterations = 100, keep_mega = FALSE, ...)

Arguments

ipm
Output from make_ipm().

...
other arguments passed to methods

n_iterations
The number of times to iterate the model to reach convergence. Default is 100.

mega_mat
A vector of names and/or 0s specifying the relationship between the kernels in the model. The names should correspond to kernel names, with 0s corresponding to sparse areas of the mega-matrix. The names can be either symbols or characters. These functions support suffix expansion as in define_k(ernel), so expressions don't need to be re-written for every combination hierarchical effects.

mega_vec
A vector of names specifying the format of the population vector. The names can be either symbols or characters.

keep_mega
A logical. TRUE returns a list with the mega matrix and full eigenvector, FALSE returns a list with each state's contribution to eigenvector. Default is FALSE, mostly for internal usage.

Value

A list of named numeric vector(s) corresponding to the stable trait distribution function (right_ev) or the reproductive values for each trait (left_ev).

Details

If the model has already been iterated, then these functions will just extract population state of the final iteration and return that in a named list. Each element of the list is a vector with length >= 1 and corresponds each state variable's portion of the eigenvector.

Note that for *_di_stoch_kern_ipm's, these generics will create a mean matrix and then compute the left/right eigenvectors for that. For *_di_stoch_param_ipm's, it will compute the average environment kernel (e.g. using the means from the env_seq slot of the IPM).

mega_mat fits the pieces of the model together to create a a transpose of the model to iterate with. Kernel names/0s should be supplied in ROW MAJOR order (think byrow = TRUE. ipmr supplies a helper function for large models format_mega_matrix to help with age-size models or ones with a lot of hierarchical effects.

Examples

data(gen_di_det_ex)

# mega matrix is specifed as a row-major vector/matrix of symbols. You can also
# supply a character vector/matrix. mega_vec is done the same way.
# DO NOT supply arguments like nrow or ncol if using \code{matrix(...)}.
# The function expects to work these out on its own.

ipmr_v <- left_ev(gen_di_det_ex,
                  mega_mat     = c(stay_discrete, go_discrete,
                                   leave_discrete, P),
                  mega_vec = c(b, ht),
                  n_iterations = 100)