Eviction correction

Various helpers to correct for unintentional eviction (Williams et al. 2012).

truncated_distributions(fun, target, ...)

discrete_extrema(target, state, ncol = NULL, nrow = NULL)

Arguments

fun

The density function to use. For example, could be "norm" to correct a Gaussian density function, or "lnorm" to correct a log-normal density function.

target

The parameter/vital rate being modified. If this is a vector, the distribution specified in fun will be recycled.

...

Used internally, do not touch!

state

The state variable used in the kernel that is being discretized.

ncol, nrow

The number of rows or column that the final form of the iteration matrix should have. This is only necessary for rectangular kernels with class "CC". make_ipm works out the correct dimensions for "DC" and "CD" kernels internally.

Value

For truncated_distributions, a modified function call with that truncates the probability density function based on the cumulative density function. For discrete_extrema, a numeric vector with modified entries based on the discretization process.

Note

Neither of these functions are intended for use outside of define_kernel, as they rely on internally generated variables to work inside of make_ipm.

References

Williams JL, Miller TEX & Ellner SP, (2012). Avoiding unintentional eviction from integral projection models.Ecology 93(9): 2008-2014.