Define implementation parameters

Define implementation parameters

define_impl(proto_ipm, kernel_impl_list)

make_impl_args_list(kernel_names, int_rule, dom_start, dom_end, ...)

Arguments

proto_ipm
The name of object you wish to append the new kernel to.

kernel_impl_list
A named list. Names correspond to kernel names. Each kernel should have 3 slots defined - the int_rule (integration rule), the dom_start (the domain the kernel begins on), and the dom_end (the domain the kernel ends on). It is safest to use make_impl_args_list to generate this.

kernel_names
A character vector with the names of the kernels that parameters are being defined for.

int_rule
The integration rule to be used for the kernel. The default is "midpoint". "trapezoid" and "g-l" (Gauss-Legendre) will be implemented as well. If "g-l", additional arguments need to be supplied (Work on this later!!).

dom_start
The name of the state variable for the kernel at time t.

dom_end
The name of the state variable for the kernel at time t+1. This is usually the same as dom_start, but general IPMs with discrete classes or IPMs that move from one state to another (e.g. tree seedling going from a height domain to a DBH domain at T+1) may have another value here. For cases with a discrete stage, kernels moving individuals from discrete to continuous should have a state variable entered here and an NA for dom_start. For kernels moving from continuous to discrete, vice versa. For discrete to discrete, both are NA.

...
Additional arguments required for int_rule = 'g-l'.

Value

A proto_ipm with the implementation details stored.

Details

define_impl is meant to help distinguish the process of generating the kernels' mathematical form from their implementation details. It takes a proto_ipm object and returns a modified one containing the implementation information.

make_impl_args_list helps generate that information in the correct format. It is usually easiest to call make_impl_args_list before calling init_ipm and then substituting that variable into the call to define_impl. Alternatively, one can do something like define_impl(kernel_impl_list = make_impl_arg_list(...)) within the course of the model definitition pipeline.