Generate an RMarkdown file with IPM metadata

Generates a .rmd file containing a mathematical description of the proto_ipm object.

make_ipm_report(
  proto_ipm,
  rmd_dest = getwd(),
  title = "",
  output_format = "html",
  render_output = FALSE,
  translate_greek = TRUE
)

make_ipm_report_body(proto_ipm, translate_greek)

Arguments

proto_ipm

A proto_ipm object

rmd_dest

The folder to save the Rmd file at. The default is getwd(). Alternatively, can be a complete file path that specifies the location and title of the document with the extension ".rmd". in this case, the current date will be appended to the title.

title

The title to include in the document. This is not necessarily the same as rmd_dest, as this appears at the top of the generated report, and is not included in the file path!

output_format

The format to include in the YAML header for the created .rmd document.

render_output

A logical indicating whether to call rmarkdown::render on the generated .rmd file. Often times, the .rmd file will need further editing before it's useful, so the default is FALSE.

translate_greek

A logical. TRUE means the function will attempt to translate symbols in vital rate expressions to Greek equivalents, which can make reports prettier. FALSE will leave the expressions largely as is before translating them to LaTeX equivalents.

Value

For make_ipm_report, the filepath to the .rmd file. The default name is "ipmr_report_<current_date>.rmd". For make_ipm_report_body, a character vector with Markdown and LaTeX suitable for rendering, but without a header.

Details

make_ipm_report_body only translates the iteration expressions and vital rate expressions into Markdown with LaTeX, and does not produce any headers needed to knit the file. This function is exported mostly for re-usage in pdb_report, and isn't really intended for use by ipmr users.

@section Translations

For iteration expressions, vital rate expressions, and parameter names, make_ipm_report first attempts to translate functional forms into a common set of names. For example, s = surv_int + surv_slope * z_1 is translated into beta_0_s + beta_1_s * z_1, and then is translated into LaTeX equations. This helps create useful Greek symbols more consistently. This may cause strange/incorrect results when parsing non-linear models.