Functions to initialize and define IPM kernels

Adds a new kernel to the proto_ipm structure.

define_kernel(
  proto_ipm,
  name,
  formula,
  family,
  ...,
  data_list = list(),
  states,
  has_hier_effs = FALSE,
  levels_hier_effs = list(),
  levels_ages = list(),
  evict_cor = FALSE,
  evict_fun = NULL,
  integrate = TRUE
)

Arguments

proto_ipm
The name of the model.

name
The name of the new kernel.

formula
A bare expression specifying the form of the kernel. See Details.

family
The type of kernel. Options are "CC" for continuous to continuous transitions, "DC" for discrete to continuous (e.g. emergence from a seedbank), "CD" for continuous to discrete (e.g. entering a seedbank), and "DD" for discrete to discrete (e.g. stasis in a seedbank).

...
For define_kernel, a set of named expressions that correspond to vital rates in formula. Suffix expansion of hierarchical models is supported.

data_list
A list of named values that correspond to constants in the formula. You do not need to specify vectors corresponding to the domains here.

states
A list with character vector containing the names of each state variable used in the kernel.

has_hier_effs
A logical indicating whether or not the kernel and/or its underlying vital rates are structured with hierarchical effects. If so and you specify either the functional forms or the exact parameter values you want to use, the kernels will be automatically split according to the notation in the vital rates and kernel formulae and multiple kernels will be built. See the vignette on the syntax for this feature for more details (vignettes( 'hierarchical-notation', package = 'ipmr')).

levels_hier_effs
A named list with vectors corresponding to the various levels the hierarchical variable can take. The names should match the suffixes used in the vital rate expressions.

levels_ages
A list with possibly 2 entries: 1. "age": the range of possible ages in the model and, optionally, 2. "max_age": the maximum age individuals in the model can attain.

evict_cor
A logical indicating whether an eviction correction should be applied to the kernel.

evict_fun
If evict_cor== TRUE, then a function that corrects for it. Currently, the only implemented function is truncated_distributions. This works by modifying the functional form of the vital rate expressions ..., and so supplying your own function in this slot will be difficult. You can also either specify a usr_funs function that performs the correction during the implementation of the model itself, or incorporate the correction into individual vital rate expressions. In either of those cases, set evict_cor to FALSE.

integrate
For simple_* models, this controls whether a "d_z" is automatically appended to the formula argument. When TRUE, this automatically creates a formula * d_z. There may be some cases where this behavior is not desirable. Set this to FALSE and specify the correct math if that happens. The default is TRUE. This argument is ignored for all general_* models.

Value

All functions described here return a proto_ipm.

Details

Different classes of IPMs may have many or only a few kernels. Each one requires its own call to this function, though there are some exceptions, namely for kernels derived from hierarchical models (e.g. vital rate models fit across plots and years). remove_k is a helper for re-building models but with a different K kernel.

Internally, ipmr generates the levels_hier_effs using expand.grid. This means that every combination of the values in levels_hier_effs must exist in the data_list, or the model will fail with an error along the lines of Error in eval_tidy: object 'x_yz' not found. In order to exclude levels that do not exist in your data, you can add a vector to the list in levels_hier_effs called "drop_levels". This should contain the values you wish to exclude as a character vector. For example, if data are collected for sites = c("a", "b", "c"), and years = c(2005:2008), but there is no data from site "a" in 2007, we can use levels_hier_effs = list(site = c("a", "b", "c"), year = c(2005:2008), drop_levels = c("a_2007")).

define_kernel generates most of the information needed to create an IPM kernel. There are a few requirements - name, family, formula, and states must not be empty. The formula should be an expression for how vital rates produce a kernel (e.g. formula = S * G). The ... should be a set of named expressions that correspond to vital rate expressions (e.g. G = dnorm(size_2, mean_size, sd_size)). See the vignettes for more information on how to get started.