Compute the per-capita growth rate for an IPM object

Compute the per-capita growth rate for a given model. Can handle stochastic and deterministic models, and has the option to discard burn in for stochastic models.

lambda(ipm, ...)

# S3 method for simple_di_det_ipm
lambda(ipm, type_lambda = "last", ...)

# S3 method for simple_di_stoch_kern_ipm
lambda(ipm, type_lambda = "stochastic", burn_in = 0.1, ...)

# S3 method for simple_di_stoch_param_ipm
lambda(ipm, type_lambda = "stochastic", burn_in = 0.1, ...)

# S3 method for general_di_det_ipm
lambda(ipm, type_lambda = "last", ...)

# S3 method for general_di_stoch_kern_ipm
lambda(ipm, ..., type_lambda = "stochastic", burn_in = 0.1)

# S3 method for general_di_stoch_param_ipm
lambda(ipm, ..., type_lambda = "stochastic", burn_in = 0.1)

# S3 method for simple_dd_det_ipm
lambda(ipm, type_lambda = "all", ...)

# S3 method for simple_dd_stoch_kern_ipm
lambda(ipm, ..., type_lambda = "stochastic", burn_in = 0.1)

# S3 method for simple_dd_stoch_param_ipm
lambda(ipm, ..., type_lambda = "stochastic", burn_in = 0.1)

# S3 method for general_dd_det_ipm
lambda(ipm, type_lambda = "last", ...)

# S3 method for general_dd_stoch_kern_ipm
lambda(ipm, ..., type_lambda = "stochastic", burn_in = 0.1)

# S3 method for general_dd_stoch_param_ipm
lambda(ipm, ..., type_lambda = "stochastic", burn_in = 0.1)

Arguments

ipm
An object returned by make_ipm().

...
other arguments passed to methods.

type_lambda
Either 'all', 'last', or 'stochastic'. 'all' returns a vector of lambda values for each time step of the simulation (equal in length to the iterations argument of make_ipm()). 'last' returns the lambda value for the final timestep. 'stochastic' returns a single value which is the geometric mean of log-per-capita growth rate from each time step.

burn_in
The proportion of iterations to discard. Default is 0.1 (i.e. first 10% of iterations in the simulation).

Value

An array. Rows correspond to time steps, and columns correspond to hierarchical effects (if any).