Print proto_ipms or *_ipm objects

Print proto_ipms or *_ipm objects

Generics for IPM classes

# S3 method for proto_ipm
print(x, ...)

# S3 method for simple_di_det_ipm
print(
  x,
  comp_lambda = TRUE,
  comp_method = "eigen",
  type_lambda = "all",
  sig_digits = 3,
  check_conv = TRUE,
  ...
)

# S3 method for simple_di_stoch_kern_ipm
print(x, comp_lambda = TRUE, type_lambda = "stochastic", sig_digits = 3, ...)

# S3 method for simple_di_stoch_param_ipm
print(x, comp_lambda = TRUE, type_lambda = "stochastic", sig_digits = 3, ...)

# S3 method for general_di_det_ipm
print(
  x,
  comp_lambda = TRUE,
  type_lambda = "last",
  sig_digits = 3,
  check_conv = TRUE,
  ...
)

# S3 method for general_di_stoch_kern_ipm
print(x, comp_lambda = TRUE, type_lambda = "stochastic", sig_digits = 3, ...)

# S3 method for general_di_stoch_param_ipm
print(x, comp_lambda = TRUE, type_lambda = "stochastic", sig_digits = 3, ...)

Arguments

x
An object produced by make_ipm.

...
Ignored

comp_lambda
A logical indicating whether or not to calculate lambdas for the iteration kernels and display them.

comp_method
Either "pop_size" or "eigen". "pop_size" computes lambda as the ratio of population sizes between successive time steps and then takes the geometric mean. "eigen" computes the dominant eigenvalue of each iteration kernel. See lambda for more details.

type_lambda
Either 'all' or 'stochastic'. See lambda for more details.

sig_digits
The number of significant digits to round to if comp_lambda = TRUE.

check_conv
A logical: for general_* models, check if population state has converged to asymptotic dynamics? If TRUE and the model has not converged, a message will be printed. Only applies to *_det when comp_method = 'pop_size'

Value

x invisibly.