Update expressions in an IPM

Provides methods to readily update vital rate functions and kernel formulae, as well as pass new parameter values and/or functions to the model.

# S3 method for proto_ipm
update(object, kernels, formulae, dots, parameters, functions, ...)

Arguments

object
A proto_ipm or output from make_ipm.

kernels
The names of the kernels to modify.

formulae
The new kernel formulae (if any). This should be a named expression or list of expressions.

dots
The new vital rate expressions (e.g. the ... of define_(k)ernel). This should be a named expression or list of expressions.

parameters
A list of new parameters to include in the model. Names that overlap with the current parameter list will overwrite the current values, while names that are not in the current parameter list will be appended to it.

functions
New user defined functions to include in the model.

...
not used