Age X Size models

Helper functions for implementing age x size models

Helper functions for implementing age x size models

# S3 method for age_size_expr
sum(expr, ..., na.rm = NA)

# S3 method for age_size_expr
prod(expr, ..., na.rm = NA)

# S3 method for age_size_expr
sum(expr, ..., na.rm = NA)

# S3 method for age_size_expr
prod(expr, ..., na.rm = NA)

Arguments

expr
An expression to compute a sum for.

...
Used internally. Modifying this will likely cause models to break.

na.rm
Not used and ignored.

Value

A modified expression.

A modified expression.

Details

Note that these functions do not compute a single number, but rather expand an expression to compute the trait distribution. This differs from the usual behavior of the numeric methods for sum, prod as those return a single number or NA.

Note that these functions do not compute a single number, but rather expand an expression to compute the trait distribution. This differs from the usual behavior of the numeric methods for sum, prod as those return a single number or NA.

Examples

if (FALSE) {

define_k(
 proto_ipm          = some_age_size_proto_ipm,
 name               = "K",
 family             = "IPM",
 n_wt_0_t_1         = sum(F_age %*% n_wt_age_t),
 n_wt_age_t_1       = P_age_minus_1 %*% n_wt_age_minus_1_t,
 n_wt_max_age_t_1   = P_max_age %*% n_wt_max_age_t +
                      P_max_age_minus_1 %*% n_wt_max_age_minus_1_t,
 data_list          = param_list,
 states             = list (c("wt")),
 has_hier_effs      = FALSE,
 levels_ages        = list(age = c(0:20), max_age = 21),
 evict_cor          = FALSE
)

}

if (FALSE) {

define_k(
 proto_ipm          = some_age_size_proto_ipm,
 name               = "K",
 family             = "IPM",
 n_wt_0_t_1         = sum(F_age %*% n_wt_age_t),
 n_wt_age_t_1       = P_age_minus_1 %*% n_wt_age_minus_1_t,
 n_wt_max_age_t_1   = P_max_age %*% n_wt_max_age_t +
                      P_max_age_minus_1 %*% n_wt_max_age_minus_1_t,
 data_list          = param_list,
 states             = list (c("wt")),
 has_hier_effs      = FALSE,
 levels_ages        = list(age = c(0:20), max_age = 21),
 evict_cor          = FALSE
)

}