Initialize an IPM

This is always the first step in constructing an IPM with ipmr. All you need for this is to know what type of IPM you want to construct - the rest comes later with define_kernel, define_k, make_ipm, and associated helper functions.

init_ipm(model_class, has_age = FALSE)

Arguments

model_class
The type of IPM. See Details for more information on classes

has_age
A logical indicating whether the model has age structure. Default is FALSE

Value

An object with classes "proto_ipm" and model_class. If has_age = TRUE, then an "age_x_size" class is also added.

Details

Combinations of simple or general, dd or di, and det or stoch are separated by an underscore to tell ipmr the type of model you will be constructing (e.g. "simple_di_det" for a deterministic IPM with a single continuous state variable and no discrete stages or density dependence).

Within stoch model types, there are two additional options to append to the class: "kern" or "param". These distinguish between models that use kernel resampling vs those that use parameter resampling (sensu Metcalf et al. 2015). Below are quick definitions, more detailed explanations can be found in the vignettes("ipmr-introduction", package = 'ipmr').

One of

simple: an IPM with a single continuous state variable that does not include any discrete stages. Simple IPMs can still be stochastic and/or density dependent.

general: an IPM with more than one continuous state variable and/or a model that includes discrete stages.

and one of

dd: used to denote a density dependent IPM.

di: used to denote a density independent IPM.

and one of

det: used to denote a deterministic IPM. Use this when you are building a single iteration kernel.

stoch: used to denote a stochastic IPM. Stochasticity can be implemented in two ways in ipmr: "kern" resampling, and "param" resampling.

If using det, this should be omitted. If using stoch, then one of the following:

kern: used to denote an IPM that uses kernel resampling. Briefly, these models build all of the iteration kernels ahead of time and then choose one at random or in a user-specified order as they move from iteration to iteration. The user-specified population vector is multiplied by the chosen kernel and the result is multiplied by the next kernel for the desired number of iterations.

param: used to denote parameter resampling. This samples distributions for each parameter based on user-specified functions supplied to define_env_state(). This will be a bit slower than "kern" resampling because iteration matrices need to be reconstructed from new parameters at every time step.

References

Metcalf et al. (2015). Statistical modelling of annual variation for inference on stochastic population dynamics using Integral Projection Models. Methods in Ecology and Evolution, 6: 1007-1017