Functions to initialize and define IPM kernels

Adds a new kernel to the proto_ipm structure. Different classes of IPMs may have many or only a few kernels. Each one requires its own call to this function, though there are some exceptions, namely for kernels derived from hierarchical models (e.g. vital rate models fit across plots and years). See Details for more info

define_kernel(
  proto_ipm,
  name,
  formula,
  family,
  ...,
  data_list = list(),
  states,
  has_hier_effs = FALSE,
  levels_hier_effs = list(),
  evict = FALSE,
  evict_fun = NULL
)

define_k(
  proto_ipm,
  name,
  family,
  ...,
  data_list = list(),
  states,
  has_hier_effs = FALSE,
  levels_hier_effs = list(),
  evict = FALSE,
  evict_fun = NULL
)

Arguments

proto_ipm
The name of object you wish to append the new kernel to.

name
The name of the new kernel - can be anything as long as it is unique to the model.

formula
A bare expression specifying the form of the kernel. See below for examples

family
The type of kernel. Options are "CC" for continuous to continuous transitions, "DC" for discrete to continuous (e.g. emergence from a seedbank), "CD" for continuous to discrete (e.g. entering a seedbank), and "DD" for discrete to discrete (e.g. stasis in a seedbank).

...
For define_kernel, set of named expressions that correspond to vital rates in formula. For define_K, a set of named expressions that relate the population state at T + 1 to the population state at T. Alternatively, can be an expression where the left hand side is the name of the kernel specified in name and the right hand side only describes the structure of the iteration kernel. In all cases, suffix expansion of hierarchical models is supported. Specific details for each case are in their respecitive function's sections below.

data_list
A list of named values that correspond to constants in the formula. You do not need to specify vectors corresponding to the domains here.

states
A character vector containing the names of each state variable used in the kernel.

has_hier_effs
A logical indicating whether or not the kernel and/or its underlying vital rates are structured with hierarchical effects. If so and you specify either the functional forms or the exact parameter values you want to use, the kernels will be automatically split according to the notation in the vital rates and kernel formulae and multiple kernels will be built. See the vignette on the syntax for this feature for more details (vignettes( 'hierarchical-notation', package = 'ipmr')).

levels_hier_effs
A named list with vectors corresponding the various levels the hierarchical variable can take. Entries in this list should be a single vector and should be character or integer typed.

evict
A logical indicating whether an eviction correction should be applied to the kernel. It is generally recommended to use a function in the vital rate definition as opposed to using this option Default is FALSE.

evict_fun
If evict == TRUE, then a function that corrects for it. Currently, the only implemented function is truncated_distributions. This works by modifying the functional form of the vital rate expressions ..., and so supplying your own function in this slot will be difficult. One can also specify usr_funs function that performs the correction during the numerical implementation of the model itself. In that case, set evict to FALSE.

Value

All functions described here return a proto_ipm.

define_k

The preferred method of defining a K kernel is to use the left hand side of the ... to reference the population vectors that the right hand side creates (e.g. n_T_1 = (P + F) * n_T). This enables powerful iteration-based methods to work properly. On the other hand, these iteration based methods can be quite time consuming, and many applications only require an iteration matrix while not necessarily requiring the population vectors (e.g. calculations of deterministic population growth rate). In those cases, the ... can contain something like K = P + F. In this case, the left hand side of the expression should match the name argument to define_k.

define_kernel

Describe vital rate expression syntax - include suffix expansion of hier_effs