Methods to implement an IPM

The make_ipm.* methods convert a proto_ipm into a set of discretized kernels and population vectors. Methods have different requirements, so be sure to read the parameter documentation. vignettes('Introduction to ipmr', 'ipmr') also contains helpful information.

make_ipm(proto_ipm, return_all = FALSE, usr_funs = list(), ...)

# S3 method for simple_di_det
make_ipm(
  proto_ipm,
  return_all = FALSE,
  usr_funs = list(),
  ...,
  domain_list = NULL,
  iterate = FALSE,
  iterations = 50,
  normalize_pop_size = TRUE
)

# S3 method for simple_di_stoch_kern
make_ipm(
  proto_ipm,
  return_all = FALSE,
  usr_funs = list(),
  ...,
  domain_list = NULL,
  iterate = FALSE,
  iterations = 50,
  kernel_seq = NULL,
  normalize_pop_size = TRUE
)

# S3 method for simple_di_stoch_param
make_ipm(
  proto_ipm,
  return_all = FALSE,
  usr_funs = list(),
  ...,
  domain_list = NULL,
  iterate = TRUE,
  iterations = 50,
  kernel_seq = NULL,
  normalize_pop_size = TRUE
)

# S3 method for general_di_det
make_ipm(
  proto_ipm,
  return_all = FALSE,
  usr_funs = list(),
  ...,
  domain_list = NULL,
  iterate = TRUE,
  iterations = 50,
  normalize_pop_size = TRUE
)

# S3 method for general_di_stoch_kern
make_ipm(
  proto_ipm,
  return_all = FALSE,
  usr_funs = list(),
  ...,
  domain_list = NULL,
  iterate = TRUE,
  iterations = 50,
  kernel_seq = NULL,
  normalize_pop_size = TRUE
)

# S3 method for general_di_stoch_param
make_ipm(
  proto_ipm,
  return_all = FALSE,
  usr_funs = list(),
  ...,
  domain_list = NULL,
  iterate = TRUE,
  iterations = 50,
  kernel_seq = NULL,
  normalize_pop_size = TRUE
)

Arguments

proto_ipm
The proto_ipm object you wish to implement. This should be the output of define_kernel, define_k, or the define_* functions.

return_all
A logical indicating whether to return the environments that the kernel expressions are evaluated in. This is useful for developer debugging and not much else.

usr_funs
An optional list of user-specified functions that are passed on to the evaluation environments. This can help make vital rate expressions more concise and expressive. Names in this list should exactly match the names of the function calls in the ... or formula.

...
Other arguments passed to methods

domain_list
An optional list of new domain information to implement the IPM with.

iterate
A logical indicating whether or not iterate the model before exiting or just return the iteration kernels. For density dependent (dd) and/or stochastic parameter resampled models (stoch_param), this should always be TRUE.

iterations
If iterate is TRUE, then the number of iterations to simulate.

normalize_pop_size
A logical indicating whether to re-scale the population vector to sum before each iteration. Default is TRUE for *_di_* methods and FALSE for *_dd_* methods.

kernel_seq
For *_stoch_kern methods, the sequence of kernels to use during the simulation process. It should have the same number of entries as the number of iterations. This can either be a vector of integers corresponding to kernel names (e.g. kernels for different years - 2011:2018), a character vector corresponding to kernel names (e.g. kernels from different sites - 'a', 'b', 'c'), a Markov chain matrix with transition probabilities between given states (NOT YET IMPLEMENTED), or empty. If it is empty, make_ipm will try to generate a sequence internally using a random selection of the levels_hier_effs defined in define_kernel.

Value

The make_ipm.* methods will always return a list of length 6 containing the following components:

iterators: iteration kernel(s) (if specified by define_k) for simple_*. otherwise contains NA.

sub_kernels: a list of sub_kernels specified in define_kernel.

env_list: a list containing the evaluation environments of kernel. This will always be empty unless return_all is TRUE. Mostly here for developer debugging.

env_seq: a matrix with dimension iterations X 1 of kernel indices indicating the order in which kernels are to be/were resampled OR a matrix with as many columns as stochastic parameters and n_iterations rows.

pop_state: population vectors stored as a list of arrays. The first dimension of each array corresponds to the state variable distribution, and the second dimsension corresponds to time steps.

proto_ipm: the proto_ipm object used to implement the model.

In addition to the list class, each object will have the class from model_class defined in init_ipm plus '_ipm'. This is to facilitate print, plot, and lambda methods. For example, a 'simple_di_stoch_kern' model will have the class 'simple_di_stoch_kern_ipm' once it has been implemented using make_ipm.

Details

When kernel_seq is a character vector, names are matched using grepl(). When it is an integer vector, the vector is first checked to make sure they are all present in kernel names. The model procedure will stop if thay are not all present.